+++ src/behaviors/ScrollShadow/DefaultContainer.js
@@ -3,5 +3,6 @@ import styled from 'styled-components';
 export default styled.div`
   position: relative;
   overflow: hidden;
-  height: 100%;
+  height: ${props => (props.horizontal ? 'auto' : '100%')};
+  width: ${props => (props.horizontal ? '100%' : 'auto')};
 `;